requirements.txt
asgiref==3.4.1
autopep8==1.5.7
click==8.0.1
dnspython==1.16.0
fastapi==0.68.0
greenlet==1.1.1
h11==0.12.0
motor==2.5.0
passlib==1.7.4
pycodestyle==2.7.0
pydantic==1.8.2
pydentic==0.0.1.dev3
PyJWT==2.2.0
python-decouple==3.4
python-dotenv==0.19.0
python-multipart==0.0.5
python-stdnum==1.16
six==1.16.0
SQLAlchemy==1.4.25
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
